pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "djvu-viewer"
description = "DjVu Viewer and tarball converter"
keywords = ["nicegui", "genealogy", "DjVu", "viewer"]
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"}
]
maintainers = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"},
]
readme = "README.md"
license= "Apache-2.0"
dependencies = [
	# https://pypi.org/project/pybasemkit/
	# https://github.com/WolfgangFahl/pybasemkit
	# Python base module kit: YAML/JSON I/O, structured logging, CLI tooling, shell execution, and remote pydevd debug support.
	"pybasemkit>=0.2.0",
	# reactive user interface
	# nicegui
	# fastapi
	# uvicorn
	#"nicegui>=2.20.0",
	"ngwidgets>=0.30.3",

	# SSO and wiki access
	# https://pypi.org/project/py-3rdparty-mediawiki/
    'py-3rdparty-mediawiki>=0.18.1',

	# CLI progress bar
	# https://pypi.org/project/tqdm/
	"tqdm>=4.66.5",

	# platform independent knowledge graph ready list of dicts (table) handling
	# https://pypi.org/project/pyLodStorage/
	"pyLodStorage>=0.18.6",

	# https://pypi.org/project/rdflib-endpoint/
	"rdflib-endpoint>=0.5.4",

	# https://github.com/FriedrichFroebel/python-djvulibre
	# https://pypi.org/project/djvulibre-python/
	"djvulibre-python>=0.9.3",

	# PNG generation
	# https://pypi.org/project/pillow/
	"pillow>=12.1.0"
]

requires-python = ">=3.10"
classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License"
    ]
dynamic = ["version"]
[tool.hatch.version]
path = "djvuviewer/__init__.py"

[project.urls]
Home = "https://github.com/WolfgangFahl/genwiki2024"
Documentation = "https://wiki.bitplan.com/index.php/genwiki2024"
Source = "https://github.com/WolfgangFahl/genwiki2024"

[project.optional-dependencies]
test = [
  "green",
]

[tool.hatch.build.targets.wheel]
only-include = ["djvuviewer","djvuviewer_examples"]

[tool.hatch.build.targets.wheel.sources]
"djvuviewer" = "djvuviewer"
"djvuviewer_examples" = "djvuviewer_examples"

[project.scripts]
djvuconv = "djvuviewer.djvu_cmd:main"
djvuviewer = "djvuviewer.djvu_webcmd:main"
djvu_migration = "djvuviewer.djvu_migration:main"


